Split repository into separate projects
