+++ build.sbt
@@ -82,9 +82,8 @@ lazy val releaseSettings = Seq(
     </scm>
     <developers>
       <developer>
-        <id>sinisa_lyh</id>
-        <name>Neville Li</name>
-        <url>https://twitter.com/sinisa_lyh</url>
+        <id>idreeskhan</id>
+        <name>Idrees Khan</name>
       </developer>
     </developers>
   }
@@ -104,20 +103,37 @@ lazy val assemblySettings = Seq(
   assemblyJarName in assembly := s"ratatool-${version.value}.jar"
 )
 
-lazy val ratatool = project
-  .settings(commonSettings)
+lazy val ratatoolCommon = project
+  .in(file("ratatool-common"))
+  .settings(commonSettings ++ noPublishSettings)
   .settings(
+    name := "ratatool-common",
     libraryDependencies ++= Seq(
-      "com.github.scopt" %% "scopt" % scoptVersion,
       "com.google.cloud.bigdataoss" % "gcs-connector" % gcsVersion,
-      "com.spotify" %% "scio-core" % scioVersion,
-      "com.spotify" %% "scio-test" % scioVersion % "test",
-      "com.twitter" %% "algebird-core" % algebirdVersion,
-      "joda-time" % "joda-time" % jodaTimeVersion,
       "org.apache.avro" % "avro" % avroVersion,
       "org.apache.avro" % "avro" % avroVersion classifier("tests"),
       "org.apache.avro" % "avro-mapred" % avroVersion classifier("hadoop2"),
       "org.apache.hadoop" % "hadoop-client" % hadoopVersion exclude ("org.slf4j", "slf4j-log4j12"),
+      "org.slf4j" % "slf4j-simple" % slf4jVersion,
+      "com.google.apis" % "google-api-services-bigquery" % bigqueryVersion,
+      "com.google.guava" % "guava" % "20.0"
+    ),
+    // In case of scalacheck failures print more info
+    testOptions in Test += Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "3")
+  )
+  .enablePlugins(ProtobufPlugin, PackPlugin)
+  .settings(protoBufSettings)
+
+lazy val ratatoolSampling = project
+  .in(file("ratatool-sampling"))
+  .settings(commonSettings)
+  .settings(
+    name := "ratatool-sampling",
+    libraryDependencies ++= Seq(
+      "com.spotify" %% "scio-core" % scioVersion,
+      "com.spotify" %% "scio-test" % scioVersion % "test",
+      "com.twitter" %% "algebird-core" % algebirdVersion,
+      "joda-time" % "joda-time" % jodaTimeVersion,
       "org.apache.parquet" % "parquet-avro" % parquetVersion,
       "org.scalacheck" %% "scalacheck" % scalaCheckVersion,
       "org.scalatest" %% "scalatest" % scalaTestVersion % "test",
@@ -127,21 +143,100 @@ lazy val ratatool = project
     testOptions in Test += Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "3")
   )
   .enablePlugins(ProtobufPlugin, PackPlugin)
-  .settings(protoBufSettings)
+  .dependsOn(
+    ratatoolCommon % "compile->compile;test->test",
+    ratatoolGenerators % "test"
+  )
 
-lazy val ratatoolScalacheck = project.in(file("ratatool-scalacheck"))
+lazy val ratatoolDiffy = project
+  .in(file("ratatool-diffy"))
+  .settings(commonSettings)
+  .settings(
+    name := "ratatool-diffy",
+    libraryDependencies ++= Seq(
+      "com.google.cloud.bigdataoss" % "gcs-connector" % gcsVersion,
+      "com.spotify" %% "scio-core" % scioVersion,
+      "com.spotify" %% "scio-test" % scioVersion % "test",
+      "com.twitter" %% "algebird-core" % algebirdVersion,
+      "joda-time" % "joda-time" % jodaTimeVersion
+    ),
+    // In case of scalacheck failures print more info
+    testOptions in Test += Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "3")
+  )
+  .enablePlugins(ProtobufPlugin, PackPlugin)
+  .dependsOn(
+    ratatoolCommon % "compile->compile;test->test",
+    ratatoolSampling,
+    ratatoolGenerators % "test"
+  )
+
+lazy val ratatoolCli = project
+  .in(file("ratatool-cli"))
+  .settings(commonSettings)
+  .settings(
+    name := "ratatool-cli",
+    libraryDependencies ++= Seq(
+      "com.github.scopt" %% "scopt" % scoptVersion,
+      "com.google.cloud.bigdataoss" % "gcs-connector" % gcsVersion,
+      "org.apache.avro" % "avro" % avroVersion,
+      "org.apache.avro" % "avro" % avroVersion classifier("tests"),
+      "org.apache.avro" % "avro-mapred" % avroVersion classifier("hadoop2"),
+      "org.apache.hadoop" % "hadoop-client" % hadoopVersion exclude ("org.slf4j", "slf4j-log4j12"),
+      "org.apache.parquet" % "parquet-avro" % parquetVersion,
+      "org.scalatest" %% "scalatest" % scalaTestVersion % "test",
+      "org.slf4j" % "slf4j-simple" % slf4jVersion
+    ),
+    // In case of scalacheck failures print more info
+    testOptions in Test += Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "3")
+  )
+  .enablePlugins(ProtobufPlugin, PackPlugin)
+  .dependsOn(
+    ratatoolCommon % "compile->compile;test->test",
+    ratatoolSampling
+  )
+
+lazy val ratatoolGenerators = project
+  .in(file("ratatool-generators"))
+  .settings(commonSettings ++ noPublishSettings)
+  .settings(
+    name := "ratatool-generators",
+    libraryDependencies ++= Seq(
+      "org.apache.avro" % "avro" % avroVersion,
+      "org.apache.avro" % "avro" % avroVersion classifier("tests"),
+      "org.apache.avro" % "avro-mapred" % avroVersion classifier("hadoop2"),
+      "org.apache.hadoop" % "hadoop-client" % hadoopVersion exclude ("org.slf4j", "slf4j-log4j12"),
+      "org.slf4j" % "slf4j-simple" % slf4jVersion,
+      "com.google.apis" % "google-api-services-bigquery" % bigqueryVersion,
+      "org.scalacheck" %% "scalacheck" % scalaCheckVersion,
+      "com.spotify" %% "scio-core" % scioVersion
+    ),
+    // In case of scalacheck failures print more info
+    testOptions in Test += Tests.Argument(TestFrameworks.ScalaCheck, "-verbosity", "3")
+  )
+  .enablePlugins(ProtobufPlugin, PackPlugin)
+  .settings(protoBufSettings)
+  .dependsOn(ratatoolCommon % "compile->compile;test->test")
+
+lazy val ratatoolScalacheck = project
+  .in(file("ratatool-scalacheck"))
   .settings(commonSettings)
   .settings(
     name := "ratatool-scalacheck",
     libraryDependencies ++= Seq(
       "org.apache.avro" % "avro" % avroVersion,
-      "org.scalacheck" %% "scalacheck" % scalaCheckVersion,
-      "com.google.apis" % "google-api-services-bigquery" % bigqueryVersion
+      "org.scalacheck" %% "scalacheck" % scalaCheckVersion
     )
-  ).dependsOn(ratatool % "test")
+  )
   .enablePlugins(ProtobufPlugin, PackPlugin)
-  .settings(protoBufSettings)
+  .dependsOn(ratatoolCommon % "compile->compile;test->test")
 
 val root = project.in(file("."))
   .settings(commonSettings ++ noPublishSettings)
-  .aggregate(ratatool, ratatoolScalacheck)
+  .aggregate(
+    ratatoolCommon,
+    ratatoolScalacheck,
+    ratatoolDiffy,
+    ratatoolSampling,
+    ratatoolCli,
+    ratatoolGenerators
+  )
