+++ ratatool-cli/src/main/scala/com/spotify/ratatool/tool/Tool.scala
