+++ ratatool-cli/src/main/scala/com/spotify/ratatool/tool/ToolConfig.scala
