+++ ratatool-cli/src/main/scala/com/spotify/ratatool/tool/ToolParser.scala
