+++ ratatool-cli/src/test/scala/com/spotify/ratatool/tool/ToolParserTest.scala
