+++ ratatool-common/src/main/avro/schema.avsc
