+++ ratatool-common/src/main/java/com/spotify/ratatool/GcsConfiguration.java
