+++ ratatool-common/src/main/protobuf/schemas.proto
