+++ ratatool-common/src/test/resources/schema.avsc
