+++ ratatool-common/src/test/resources/schema.json
