+++ ratatool-common/src/test/scala/com/spotify/ratatool/Schemas.scala
