+++ ratatool-diffy/src/main/scala/com/spotify/ratatool/diffy/ProtoBufDiffy.scala
