+++ ratatool-diffy/src/test/scala/com/spotify/ratatool/diffy/AvroDiffyTest.scala
