+++ ratatool-diffy/src/test/scala/com/spotify/ratatool/diffy/ProtoBufDiffyTest.scala
