+++ ratatool-generators/src/main/scala/com/spotify/ratatool/generators/AvroGenerator.scala
