+++ ratatool-generators/src/main/scala/com/spotify/ratatool/generators/ProtoBufGenerator.scala
