+++ ratatool-generators/src/test/scala/com/spotify/ratatool/scalacheck/AvroGenSpec.scala
