+++ ratatool-sampling/src/main/scala/com/spotify/ratatool/io/TableRowJsonIO.scala
