+++ ratatool-sampling/src/test/scala/com/spotify/ratatool/io/AvroIOTest.scala
